% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{deleteBucket}
\alias{deleteBucket}
\title{Delete an App-Managed Bucket.}
\usage{
deleteBucket(token = NULL, bucket = "mybucket")
}
\arguments{
\item{token}{A string or \code{aps_token} object with
\code{bucket:delete} scope.}

\item{bucket}{A string. Name of the bucket to delete. Defaults to
\code{mybucket}.}
}
\value{
An object containing the HTTP status code of the response.
}
\description{
Delete an app-managed bucket using the Data Management API.
}
\examples{
\dontrun{
# Delete a bucket named "mybucket"
resp <- deleteBucket(token = myToken, bucket = "mybucket")
}
}
