% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{deleteObject}
\alias{deleteObject}
\title{Delete an Object from an App-Managed Bucket.}
\usage{
deleteObject(token = NULL, bucket = "mybucket", object = NULL)
}
\arguments{
\item{token}{A string or \code{aps_token} object with
\code{data:write} scope.}

\item{bucket}{A string. Name of the bucket. Defaults to \code{mybucket}.}

\item{object}{A string. Key (name) of the object to delete.}
}
\value{
An object containing the HTTP status code of the response.
}
\description{
Delete an object from an app-managed bucket using the Data Management API.
}
\examples{
\dontrun{
# Delete the "aerial.dwg" object from "mybucket"
resp <- deleteObject(token = myToken, bucket = "mybucket", object = "aerial.dwg")
}
}
