% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_expired}
\alias{is_expired}
\title{Check Whether an aps_token Has Expired.}
\usage{
is_expired(token)
}
\arguments{
\item{token}{An \code{aps_token} object returned by \code{\link{getToken}}.}
}
\value{
Logical. \code{TRUE} if the token has expired, \code{FALSE} otherwise.
}
\description{
Check Whether an aps_token Has Expired.
}
\examples{
\dontrun{
tok <- getToken(id = Sys.getenv("client_id"), secret = Sys.getenv("client_secret"))
is_expired(tok)
}
}
