% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{listBuckets}
\alias{listBuckets}
\title{List All App-Managed Buckets.}
\usage{
listBuckets(token = NULL, limit = 10, startAt = NULL, region = "US")
}
\arguments{
\item{token}{A string or \code{aps_token} object with
\code{bucket:read} scope.}

\item{limit}{An integer. Maximum number of buckets to return. Defaults to
\code{10}.}

\item{startAt}{A string. Bucket key to start the list from (for pagination).
Defaults to \code{NULL}.}

\item{region}{A string. Region filter. May be \code{"US"} or \code{"EMEA"}.
Defaults to \code{"US"}.}
}
\value{
An object containing a list of bucket details.
}
\description{
List all app-managed buckets using the Data Management API.
}
\examples{
\dontrun{
# List all buckets
resp <- listBuckets(token = myToken)
resp$content$items
}
}
