% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realityCapture.R
\name{processPhotoscene}
\alias{processPhotoscene}
\title{Start Reality Capture Processing.}
\usage{
processPhotoscene(photoscene_id = NULL, token = NULL)
}
\arguments{
\item{photoscene_id}{A string. Photoscene ID returned by
\code{\link{createPhotoscene}}.}

\item{token}{A string or \code{aps_token} object with
\code{data:read} and \code{data:write} scopes.}
}
\value{
An object of class \code{processPhotoscene} containing the
  processing response.
}
\description{
Initiate photogrammetry processing for a photoscene that has had images
uploaded via \code{\link{uploadImages}}.
}
\examples{
\dontrun{
proc <- processPhotoscene(photoscene_id = myPhotosceneId, token = myToken)
}
}
