% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDerivative.R
\name{checkFile}
\alias{checkFile}
\title{Check the Status of a Translated File.}
\usage{
checkFile(urn = NULL, token = NULL)
}
\arguments{
\item{urn}{A string. Source URN (objectId) for the file. Note the URN must be
Base64 encoded. To encode the URN, see, for example, the
\code{jsonlite::base64_enc} function.}

\item{token}{A string. Token generated with \code{\link{getToken}} function
with \code{data:read} and \code{data:write} scopes.}
}
\description{
Check the status of a recently translated file using the Model Derivative API.
}
\examples{
\dontrun{
# Check the status of the translated "aerial.dwg" svf file
myUrn <- jsonlite::base64_enc(Sys.getenv("urn"))
checkFile(urn <- myUrn, token = Sys.getenv("access_token"))
}
}
\seealso{
\url{https://developer.autodesk.com/en/docs/model-derivative/v2/overview/}
}

