% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManagement.R
\name{listObjects}
\alias{listObjects}
\title{List Objects in an App-Managed Bucket.}
\usage{
listObjects(token = NULL, bucket = "mybucket", limit = 10)
}
\arguments{
\item{token}{A string. Token generated with \code{\link{getToken}} function
with \code{data:read} scope.}

\item{bucket}{A string. Name of the bucket. Defaults to \code{mybucket}.}

\item{limit}{An integer. Maximum number of objects to return. Defaults to
\code{10}.}
}
\value{
An object containing a list of objects in the bucket.
}
\description{
List objects stored in an app-managed bucket using the Data Management API.
}
\examples{
\dontrun{
# List objects in "mybucket"
resp <- listObjects(token = myToken, bucket = "mybucket")
resp$content$items
}
}
