% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDerivative.R
\name{translateObj}
\alias{translateObj}
\title{Translate a File into OBJ Format.}
\usage{
translateObj(urn = NULL, token = NULL)
}
\arguments{
\item{urn}{A string. Source URN (objectId) for the file. Note the URN must be
Base64 encoded. To encode the URN, see, for example, the
\code{jsonlite::base64_enc} function.}

\item{token}{A string. Token generated with \code{\link{getToken}} function
with \code{data:read} and \code{data:write} scopes.}
}
\value{
An object containing the \code{result}, \code{urn}, and additional
  activity information.
}
\description{
Translate a File into OBJ Format.
}
\examples{
\dontrun{
# Translate the "aerial.dwg" file into a obj file
myUrn <- jsonlite::base64_enc(Sys.getenv("urn"))
translateObj(urn <- myUrn, token = Sys.getenv("token"))
}
}
\seealso{
\url{https://developer.autodesk.com/en/docs/model-derivative/v2/overview/}
}

